FreeRTOS: Add support for SMP
When only a single core is used, FreeRTOS stores the currently
executing task in pxCurrentTCB but when multiple cores are running
it stores it in an array pxCurrentTCBs.  Use pxCurrentTCBs[0] if
in a multi-core environment and fall back to pxCurrentTCB in a single
core environment.

The location of the task name in the TCB structure also changes
based on whatever options are enabled in the SMP code.  Use
(a currently unmerged commit to FreeRTOS that adds) uxTaskNameOffset
which contains the offset within the TCB of the task name.  If this
doesn't exist, fall back to the previous definition of 52 bytes.